Helper class for computing finance data.
The point of this class is to gather all of the finance data outside
of the template, so it can be used in a CSV (or anything else).  With
some minor adjustments it could be used to get the data that is used
in the `finances.html` template, but for right now they are
duplicating each other.