Remove Order lookup by code.
An order code in the url is no longer a valid way to retrieve order
details.  If someone makes a request to a url with the code, that
request will either be redirected to the event's order page without
the code.  The system will then lookup the logged-in-user's order
information, or retrieve it from the session if that user is not
logged in.

There's one edge case here, and that's where a non-logged-in user is
looking up an extant unclaimed order by its code and there's nothing
in the session to let us know the order belongs to the user.  This
might happen if someone orders something then clicks the url in a
different browser.  In this case we redirect the user to the login
page where they can either log in to claim the order, or create an
account to claim the order.