Stopped from sending requests if we know it will fail. Fixed #360.
Item buttons in the shop are given the class "disabled" if they are
sold out, so if something is disabled, don't even bother trying to add
it to the user's shopping cart.  Clicking on these items now does
nothing at all, which is in line with how most disabled form elements
work.