Activated OCSP stapling.
Fetches OCSP records and caches them, which will convey
certificate-revocation information to visitors. See:

 - http://blog.mozilla.org/security/2013/07/29/ocsp-stapling-in-firefox/
 - https://matthiasadler.info/blog/ocsp-stapling-on-nginx-with-comodo-ssl/

Requires a file containing trusted CA certificates to verify the OCSP
response.  This is the same as the certificate chain we are using for
dancerfly.com itself, minus the actual certificate for our server.

Adds a `resolver` line to increase the speed of OCSP verification
requests.