Changed liability waiver to use Template.
This seems better thant `string.format()` because (1) we are only using
basic substitutions and (2) we can make those substitutions more safely.

Currently, this is backwards compatible with the existing waivers in
the database that use `{bracketed}` syntax.  However, we have to convert
it to `${dollars bracket}` or just `$dollars` syntax to use the Template
class.  That's kind of a shim, so in the future we might want to:

1. Convert existing waivers in the DB to use `$event` and
`$organization`.
2. Change the help text for this field to reflect the above.

But that's not urgent.