Added webhook for processing stripe refunds.
Basic overview:

1. We receive event from stripe.
2. We only trust 2 fields from the initial payload: `id` and
`livemode`.  We query for the rest of the fields from Stripe directly
using that ID and the key corresponding to live or test accordingly.
3. We currently only process `charge.refunded` event types.
4. If this event has already been seen, ignore it.
5. Query additional data from stripe: refund information, charge
information with balance transactions expanded, application fees.
6. Create a refund transaction on our own system.

Note about return status codes: Stripe will attempt to resend the
request unless it receives a `2xx` response code.  So I've attempted to
separate error conditions into "We will never process this, do not
resend" (meaning `200`) or "Something seems like it was wrong with the
request itself, try again" (meaning `404`).