Refactor webhook tests
So, I changed around a bunch of the tests and I think they're better
for it.  Mostly I tried to _just_ mock the initial Event.retrieve
call, since our event has to be fake, and _actually_ have everything
else come from the test API, which is recorded with VCR.

This actually revealed some strangeness with the Application Fees,
which I've fixed and am pretty confident that this webhook view does
the right thing.