Remove unnecessary vcr recordings
At some point in the lifecycle of this code, I was using vcr more
extensively than I am now, and these recordings are not necessary
anymore.  We are mocking the Event.retrive API call to raise an error,
which causes our webhook function to return early and make no more
calls.  So the VCR is not necessary since no other calls even happen.